+++ changelog.md
@@ -4,7 +4,8 @@ Todas las modificaciones notables a este proyecto se documentarán en este archi
 
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/es/1.0.0/).
 
-## [1.1.2] - 2025-09-15
+## [1.1.3] - 2025-09-15
 
 ### Agregado 
-- Metodo de mover fichas
+-Test mover ficha 
+-Inicio de turnos